New deploy done. Ready for first test for initial deploy

+++ deployment/steps.py
@@ -1,8 +1,8 @@
 import os
+import time
 import diff
 from docker import Client
 from dockerutils import container_utils, image_utils, misc_utils
-import params
 import shutil
 import sys
 
@@ -38,7 +38,7 @@ def verify_changes(base_dir, logger):
 
 def create_images(deploy_settings):
     # We can always update all images to their new versions. 
-    images_to_create = ['postgres', 'mongo','webapp','nginx']
+    images_to_create = ['postgres', 'mongo','webapp','nginx','upgradenginx']
 
     # if deploy_settings['deploy_extent'] in ['ALL', 'CALCULATIONS'] and deploy_settings['files_changed']:
 
@@ -137,7 +137,7 @@ def run_calculations(deploy_settings):
 
 def stop_and_remove_current_containers(deploy_settings):
     
-    containers_to_stop = ['nginx', 'webapp', 'mongo', 'postgres']
+    containers_to_stop = ['nginx', 'webapp', 'mongo', 'postgres', 'webapp-data', 'mongo-data', 'postgres-data']
 
     for container in containers_to_stop:
         if container_utils.isContainerRunning(container):
@@ -145,39 +145,34 @@ def stop_and_remove_current_containers(deploy_settings):
             cli.stop(container)
         if container_utils.isContainerPresent(container):
             deploy_settings['log'].info("Removing " + container)
-            cli.remove_container(container)
+            cli.remove_container(container, v=True)
 
-def create_and_start_data_containers(deploy_settings):
-
+def create_and_start_data_containers(deploy_settings, tag='latest'):
     containers_to_start = ['webapp', 'postgres', 'mongo']
 
     for container in containers_to_start:
         if container_utils.isContainerRunning(container+'-data') == False:
             deploy_settings['log'].info("Creating and starting data container: {0}".format(container+'-data'))
-            cont = cli.create_container(image='demokratikollen/'+ container + ':latest',
+            cont = cli.create_container(image='demokratikollen/'+ container + ':' + tag,
                                         name=container+'-data',
                                         command='/bin/sh -c "while true; do sleep 1; done"')
-            cli.start(cont['Id'])
+            cli.start(cont['Id'], restart_policy = {"MaximumRetryCount": 0, "Name": 'always'})
 
-
-def create_and_start_app_containers(deploy_settings):
+def create_and_start_app_containers(deploy_settings, tag='latest'):
+    restart_policy = {"MaximumRetryCount": 0, "Name": 'always'}
     deploy_settings['log'].info("Creating and starting postgres container.")
-    cont = cli.create_container(image='demokratikollen/postgres:latest', name='postgres')
-    cli.start(cont['Id'], volumes_from='postgres-data')
+    cont = cli.create_container(image='demokratikollen/postgres:'+tag, name='postgres')
+    cli.start(cont['Id'], volumes_from='postgres-data', restart_policy=restart_policy)
 
     deploy_settings['log'].info("Creating and starting mongo container.")
-    cont = cli.create_container(image='demokratikollen/mongo:latest', name='mongo')
-    cli.start(cont['Id'], volumes_from='mongo-data')
+    cont = cli.create_container(image='demokratikollen/mongo:'+tag, name='mongo')
+    cli.start(cont['Id'], volumes_from='mongo-data',restart_policy=restart_policy)
 
     deploy_settings['log'].info("Creating and starting webapp container.")
-    cont = cli.create_container(image='demokratikollen/webapp:latest', name='webapp')
-    cli.start(cont['Id'],volumes_from='webapp-data',links={'postgres': 'postgres', 'mongo':'mongo'})
+    cont = cli.create_container(image='demokratikollen/webapp:'+tag, name='webapp')
+    cli.start(cont['Id'],volumes_from='webapp-data',links={'postgres': 'postgres', 'mongo':'mongo'},restart_policy=restart_policy)
 
-    deploy_settings['log'].info("Creating and starting nginx container.")
-    cont = cli.create_container(image='demokratikollen/nginx:latest', name='nginx')
-    cli.start(cont['Id'],volumes_from='webapp-data', links={'webapp': 'webapp'}, port_bindings={80:80})
-
-def update_database_data(deploy_settings):
+def update_database_data(deploy_settings, tag='latest'):
     data_dir = os.path.join(deploy_settings['base_dir'],'data/database_dumps')
     binds = {}
     binds[data_dir] = {'bind': '/data', 'ro': False}
@@ -186,33 +181,47 @@ def update_database_data(deploy_settings):
     links['postgres'] = 'postgres'
     links['mongo'] = 'mongo'
 
+    if tag == 'latest' and os.path.isfile(os.path.join(data_dir, 'demokratikollen_postgres_latest.gz')):
+        postgres_path = '/data/demokratikollen_postgres_latest.gz'
+    else:
+        postgres_path = '/data/demokratikollen_postgres_current.gz'
+
+    if tag == 'latest' and os.path.isdir(os.path.join(data_dir, 'demokratikollen_mongo_latest')):
+        mongo_path = '/data/demokratikollen_mongo_latest'
+    else:
+        mongo_path = '/data/demokratikollen_mongo_current'     
+
     deploy_settings['log'].info("Updating the Postgres database.")
-    cont = cli.create_container(image='demokratikollen/postgres:latest',
+    cont = cli.create_container(image='demokratikollen/postgres:' + tag,
                                 command='/bin/sh -c "while true; do sleep 1; done"',
                                 volumes='/data')
     cli.start(cont['Id'], binds=binds, links=links)
     container_utils.run_command_in_container(cont['Id'],
-        "/bin/bash -c 'gunzip -c /data/demokratikollen_postgres_latest.gz | psql -h postgres -U demokratikollen demokratikollen'",
+        "/bin/bash -c 'gunzip -c " + postgres_path + " | psql -h postgres -U demokratikollen demokratikollen'",
         deploy_settings['log'])
     cli.remove_container(cont['Id'], v=True, force=True)
 
     deploy_settings['log'].info("Updating the mongo database.")
-    cont = cli.create_container(image='demokratikollen/mongo:latest',
+    cont = cli.create_container(image='demokratikollen/mongo:' + tag,
                                 command='/bin/sh -c "while true; do sleep 1; done"',
                                 volumes='/data')
     cli.start(cont['Id'], binds=binds, links=links)
     container_utils.run_command_in_container(cont['Id'],
-        "mongorestore --host mongo --drop /data/demokratikollen_mongo_latest",
+        "mongorestore --host mongo --drop "+ mongo_path,
         deploy_settings['log'])
     cli.remove_container(cont['Id'], v=True, force=True)    
 
-def update_webapp_src(deploy_settings):
+def update_webapp_src(deploy_settings, tag='latest'):
 
-    web_src_dir = os.path.join(deploy_settings['base_dir'],'demokratikollen')
+    if tag == 'latest':
+        web_src_dir = os.path.join(deploy_settings['base_dir'],'demokratikollen')
+    if tag == 'current':
+        web_src_dir = os.path.join(deploy_settings['base_dir'],'demokratikollen_old')
+
     binds = {}
     binds[web_src_dir] = {'bind': '/mnt/websrc', 'ro': True}
 
-    cont = cli.create_container(image='demokratikollen/webapp:latest',
+    cont = cli.create_container(image='demokratikollen/webapp:'+tag,
                         name='webapp-temp',
                         command='/bin/sh -c "while true; do sleep 1; done"',
                         volumes='/mnt/websrc')
@@ -239,7 +248,13 @@ def post_deployment(deploy_settings):
                    'demokratikollen/webapp']
     for image_name in image_names:
         cli.tag(image=image_name, tag='current', repository=image_name, force=True)
+
+    deploy_settings['log'].info("Removing any untagged images.")
+    image_utils.removeUntaggedImages()
     
+    deploy_settings['log'].info("Removing any orphaned volumes")
+    misc_utils.removeUnusedVolumes()
+
     if deploy_settings['redo_calculations']:
         current_mongo_database_dump =  os.path.join(deploy_settings['base_dir'],'data/database_dumps/demokratikollen_mongo_current')
         latest_mongo_database_dump  =  os.path.join(deploy_settings['base_dir'],'data/database_dumps/demokratikollen_mongo_latest')
@@ -248,17 +263,74 @@ def post_deployment(deploy_settings):
 
         deploy_settings['log'].info("Saving latest database dumps")
         
+        if not os.path.isdir(latest_mongo_database_dump):
+            raise Exception
+        if not os.path.isfile(latest_postgres_database_dump):
+            raise Exception
+
         shutil.rmtree(current_mongo_database_dump)
         shutil.move(latest_mongo_database_dump, current_mongo_database_dump)
 
         os.remove(current_postgres_database_dump)
         shutil.move(latest_postgres_database_dump, current_postgres_database_dump)
 
+def start_upgrade_message(deploy_settings):
+    if container_utils.isContainerRunning('nginx'):
+        cli.stop('nginx')
+
+    if container_utils.isContainerPresent('upgradenginx'):
+        cli.remove_container('upgradenginx', force=True, v=True)
+    cont = cli.create_container(image='demokratikollen/upgradenginx', name='upgradenginx')
+    cli.start(cont['Id'], port_bindings={80:80})
+
+def stop_upgrade_message(deploy_settings, tag='latest'):
+    cli.stop('upgradenginx')
+    cli.remove_container('upgradenginx')
+
+    cont = cli.create_container(image='demokratikollen/nginx:'+tag, name='nginx')
+    cli.start(cont['Id'],
+        volumes_from='webapp-data',
+        links={'webapp': 'webapp'}, 
+        port_bindings={80:80},
+        restart_policy={'MaximumRetryCount': 5, 'Name': 'always'} )
+
 def clean_up_after_error(deploy_settings):
+    deploy_settings['log'].info("Removing all running containers")
+    container_utils.remove_all_containers()
+
+    deploy_settings['log'].info("Starting upgrading message.")
+    cont = cli.create_container(image='demokratikollen/upgradenginx', name='upgradenginx')
+    cli.start(cont['Id'], port_bindings={80:80})    
+
+    deploy_settings['log'].info("Runnning last deploy")
+    create_and_start_data_containers(deploy_settings, tag='current')
+    update_webapp_src(deploy_settings, tag='current')
+    create_and_start_app_containers(deploy_settings, tag='current')
+
+    deploy_settings['log'].info("Waiting 30 seconds for services to boot")
+    time.sleep(30)
+
+    update_database_data(deploy_settings, tag='current')
+    stop_upgrade_message(deploy_settings, tag='current')
+
     deploy_settings['log'].info("Removing current src")
-    shutil.rmtree(os.path.join(deploy_settings['base_dir'],'demokratikollen'))
+    src_path = os.path.join(deploy_settings['base_dir'],'demokratikollen')
+    if os.path.isdir(src_path):
+        shutil.rmtree(src_path)
+
     deploy_settings['log'].info("Cleaning up untagged Images")
     image_utils.removeUntaggedImages(force=True)
+
+    deploy_settings['log'].info("Removing any orphaned volumes")
+    misc_utils.removeUnusedVolumes()
+    
+    deploy_settings['log'].info("Removing latest databasedumps")
+    latest_mongo_database_dump  =  os.path.join(deploy_settings['base_dir'],'data/database_dumps/demokratikollen_mongo_latest')
+    latest_postgres_database_dump  =  os.path.join(deploy_settings['base_dir'],'data/database_dumps/demokratikollen_postgres_latest.gz')
+    if os.path.isdir(latest_mongo_database_dump):
+        shutil.rmtree(latest_mongo_database_dump)
+    if os.path.isfile(latest_postgres_database_dump):
+        os.remove(latest_postgres_database_dump)
 
 def create_image(name,deploy_settings):
     cli = Client(base_url='unix://var/run/docker.sock')
@@ -269,10 +341,13 @@ def create_image(name,deploy_settings):
     deploy_settings['log'].info("Creating image: {0}".format(name))
     for line in cli.build(path=docker_path, rm=True, tag=full_name):
         log_data = misc_utils.decode_docker_log(line)
-
         if "error" in log_data:
             deploy_settings['log'].error("Something went wrong with docker: {0}".format(log_data['error']))
             raise Exception
         else:
-            deploy_settings['log'].info("Docker: {0}".format(log_data['stream'].strip()))
+            if 'stream' in log_data:
+                msg = log_data['stream'].strip()
+            if 'status' in log_data:
+                msg = log_data['status'].strip()
+            deploy_settings['log'].debug("Docker: {0}".format(msg))
